calculator
float type added

+++ resources/views/admin/master.blade.php
@@ -133,8 +133,8 @@ function myFunction() {
 <script>
 
 function calcular(){
-    var valor1 = parseInt(document.getElementById('n1').value,10);
-    var valor2 = parseInt(document.getElementById('n2').value,10);
+    var valor1 = parseFloat(document.getElementById('n1').value,10);
+    var valor2 = parseFloat(document.getElementById('n2').value,10);
     document.getElementById('result').value = valor1 * valor2;
 }
 </script>
